ChatbotOnSlack/requirement.txt
aiohttp==3.6.2
altgraph==0.17
async-timeout==3.0.1
attrs==19.3.0
boto3==1.12.36
botocore==1.15.36
cachetools==4.0.0
CD==1.1.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
docutils==0.15.2
easygui==0.98.1
et-xmlfile==1.0.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SocketIO==4.2.1
future==0.18.2
gensim==3.8.1
get==2019.4.13
google-api-core==1.16.0
google-auth==1.13.1
google-cloud-core==1.3.0
google-cloud-storage==1.27.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.9
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
MarkupSafe==1.1.1
multidict==4.7.5
numpy==1.16.2
openpyxl==3.0.3
pandas==0.25.3
pefile==2019.4.18
Pillow==7.1.1
post==2019.4.13
protobuf==3.11.3
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==7.0.1
PyInstaller==3.6
python-dateutil==2.8.1
python-engineio==3.12.1
python-socketio==4.5.1
pytz==2019.3
pywin32-ctypes==0.2.0
query-string==2019.4.13
reportlab==3.5.42
request==2019.4.13
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
scipy==1.4.1
six==1.14.0
slackcli==1.0.0
slackclient==1.0.5
slackclient-cli==0.4.6
slackeventsapi==2.1.0
smart-open==1.10.0
urllib3==1.25.8
websocket-client==0.57.0
Werkzeug==1.0.1
XlsxWriter==1.2.8
yarl==1.4.2
